* stage/fix-warnings: (43 commits)
  Fix conversion warning in vtkSICompoundSourceProxy
  Fix unused variable warning in vtkSpyPlotReader
  Disable assignment operator for vtkSpyPlotIStream
  Fix conversion warning in vtkHierarchicalFractal
  Fix conversion warnings in vtkGridConnectivity
  Disable assignment operator for vtkRecordMTime
  Fix variable conversion warning in vtkExtractHistogram
  Fix unitialized variable warning in vtkAMRDualGridHelper
  Disable assignment operator for vtkPVSystemInformationSave and Load
  Disable assignment operator for EdgeEndpoints
  Initialize variables in vtkMultiSliceContextItem
  Use bool instead of int for visibility flag in vtkChartNamedOptions
  Fix ignored warning regex for protobuf
  Remove unused variable in pqPropertiesPanel
  Fixing static tests
  Add missing test properties for static tests
  Auto reset slices origin to first representation bound center
  Add QuadView testing for static dashboard
  Properly fix TileDisplay rendering
  Update for Composite Iterator API change
  ...

+++ ParaViewCore/VTKExtensions/Default/vtkHierarchicalFractal.cxx
@@ -129,7 +129,7 @@ public:
       }
 
     std::vector<unsigned int> blockIds(blocksPerLevel.size(),0); //keep track of the id at each level
-    hbds->Initialize(blocksPerLevel.size(), &blocksPerLevel[0]);
+    hbds->Initialize(static_cast<int>(blocksPerLevel.size()), &blocksPerLevel[0]);
     hbds->SetOrigin(origin);
     hbds->SetGridDescription(this->GridDescription);
     for(unsigned int level=0; level<hbds->GetNumberOfLevels(); level++)
